part 1
fix

done

address comments
